+++ test/UrlModifierTest.php
@@ -81,16 +81,25 @@ class UrlModifierTest extends PHPUnit_Framework_TestCase
         $this->assertSame('example.com', $url->getHost());
     }
 
-
     public function testWithoutZoneIdentifier()
     {
         $url = Url::createFromString(
             'http://[fe80::1234%25eth0-1]/path/../to/the/./sky.php?kingkong=toto&foo=bar+baz#doc3'
         );
-        $url = $url->withoutZoneIdentifier();
-        $this->assertSame('[fe80::1234]', $url->getHost());
+        $this->assertSame('[fe80::1234]', $url->withoutZoneIdentifier()->getHost());
     }
 
+    public function testToUnicode()
+    {
+        $url = Url::createFromString('ftp://xn--mgbh0fb.xn--kgbechtv/where/to/go');
+        $this->assertSame('مثال.إختبار', $url->toUnicode()->getHost());
+    }
+
+    public function testToAscii()
+    {
+        $url = Url::createFromString('ftp://مثال.إختبار/where/to/go');
+        $this->assertSame('xn--mgbh0fb.xn--kgbechtv', $url->toAscii()->getHost());
+    }
 
     public function testFilterQueryParameters()
     {
@@ -168,5 +177,17 @@ class UrlModifierTest extends PHPUnit_Framework_TestCase
     {
         $url = $this->url->withExtension('asp');
         $this->assertSame('/path/to/the/sky.asp', (string) $url->getPath());
+    }
+
+    public function testWithTrailingSlash()
+    {
+        $url = $this->url->withTrailingSlash();
+        $this->assertSame('/path/to/the/sky.php/', (string) $url->getPath());
+    }
+
+    public function testWithoutTrailingSlash()
+    {
+        $url = Url::createFromString('http://www.example.com/');
+        $this->assertSame('', (string) $url->withoutTrailingSlash()->getPath());
     }
 }